backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
earthengine-api==0.1.388
google-generativeai==0.3.2
pydantic==2.6.0
python-dotenv==1.0.1
numpy==1.26.3
geojson==3.1.0
torch==2.2.0
scikit-learn==1.4.0
joblib==1.3.2
matplotlib==3.8.2
Pillow==10.2.0
rasterio==1.3.9
python-pptx==0.6.23
